+++ artifacts/helm/http_values.yaml
@@ -23,27 +23,23 @@ deployment:
     MODE: http
     LOAD_RUN_ID: "" # if empty, a random UUID will be generated
     # CONCURRENCY determines how many slots are used to send data to the server.
-    # Each slot is a unique pair of "source <-> user_id".
-    # CONCURRENCY should always be a multiple of 100 and be greater or equal to 100.
     CONCURRENCY: "200"
     MESSAGE_GENERATORS: "10"
     # SOURCES should be a comma separated list of writeKeys
     # e.g. SOURCES: "2lNXnjJU9xrbUERT3Uy3Po8jKbr,2nYfF7hsD7KXz0Vp4SW1TivZCRu"
     # This goes together with the number of replicas. You'll need one source per replica here.
     SOURCES: "2lNXnjJU9xrbUERT3Uy3Po8jKbr,2nYfF7hsD7KXz0Vp4SW1TivZCRu"
-    RANDOM_KEY_NAMES: "false" # random userIDs, use false if you want the user_ids to be predictable e.g. user_1, user_2...
-    # Keys per slot map e.g. 50_30_20 means the 1st group gets 50 keys each, the 2nd 30 and the 3rd 20.
-    # If single value then everybody gets the same number of keys.
-    KEYS_PER_SLOT_MAP: "50"
-    # Traffic distribution e.g. 50_30_20 means the 1st group gets 50% of the traffic, the 2nd 30% and the 3rd 20%.
-    # If single value then everybody gets the same percentage.
-    TRAFFIC_DISTRIBUTION_PERCENTAGE: "60_40"
     USE_ONE_CLIENT_PER_SLOT: "true"
     ENABLE_SOFT_MEMORY_LIMIT: "true"
-    TOTAL_MESSAGES: "1000000"
-    TOTAL_DURATION: "0" # 0 => finish as fast as possible
-    RUDDER_EVENTS_BATCH_SIZE: 2 # do 0 to send single "page" event
-    HTTP_ENDPOINT: "http://myns-ingestion.myns:8080/v1/batch"
+    TOTAL_USERS: "100000"
+    # HOT_USER_GROUPS: sum should be 100 (%) and values comma separated
+    # TOTAL_USERS will be divided by the number of groups and given the desired data concentration
+    HOT_USER_GROUPS: "30,20,10,10,10,10,5,5"
+    EVENT_TYPES: "page,batch(10,0),batch(30,100)"
+    # HOT_EVENT_TYPES: sum should be 100 (%) and values comma separated
+    # It should be a 1:1 match with the groups in EVENT_TYPES.
+    # The groups here define the percentage of the events in EVENT_TYPES.
+    HOT_EVENT_TYPES: "33,33,34"
     HTTP_COMPRESSION: "true"
     HTTP_READ_TIMEOUT: "5s"
     HTTP_WRITE_TIMEOUT: "5s"
@@ -51,3 +47,4 @@ deployment:
     HTTP_MAX_CONNS_PER_HOST: "5000"
     HTTP_CONCURRENCY: "1000"
     HTTP_CONTENT_TYPE: "application/json"
+    HTTP_ENDPOINT: "http://myns-ingestion.myns:8080/v1/batch"
